[김정우] 메인에서의 아티팩트 구현

+++ Assets/Scripts/Data/Enemy/EnemyData.cs
@@ -41,16 +41,16 @@ public class EnemyData
 
     public int Id;
     public string Name;
-    public long Attack;
-    public long MagicalAttack;
-    public long Health;
-    public long Defence;
-    public long MagicDefence;
-    public long Shield;
-    public long Accuracy;
-    public long Evasion;
-    public long Critical;
-    public long AttackSpeed;
+    public int Attack;
+    public int MagicalAttack;
+    public int Health;
+    public int Defence;
+    public int MagicDefence;
+    public int Shield;
+    public int Accuracy;
+    public int Evasion;
+    public int Critical;
+    public int AttackSpeed;
     public Sprite Image;
     public RuntimeAnimatorController RuntimeAnimatorController;
 }
